requirements.txt
# Python 依赖包
# @author ygw

# 核心依赖
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Lean 相关
lean-dojo>=1.0.0

# 检索相关
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 训练相关
trl>=0.12.0
peft>=0.13.0
wandb>=0.15.0
tensorboard>=2.13.0
bitsandbytes>=0.43.0

# 工具库
tqdm>=4.65.0
loguru>=0.7.0
